stable

+++ SteamRoot/Zero-K_linux64.sh
@@ -1,6 +1,12 @@
 #!/bin/sh
 
-ln -fs /lib/x86_64-linux-gnu/libc.so.6 linux64/libc.so
+if [ -f /lib/x86_64-linux-gnu/libc.so.6 ]; then
+    ln -fs /lib/x86_64-linux-gnu/libc.so.6 linux64/libc.so
+elif [ -f /lib64/libc.so.6 ]; then
+    ln -fs /lib64/libc.so.6 linux64/libc.so
+elif [ -f /lib/libc.so.6 ]; then
+    ln -fs /lib/libc.so.6 linux64/libc.so
+fi
 
 export LD_LIBRARY_PATH="./linux64:$LD_LIBRARY_PATH"
 ./Zero-K_linux64 "$@"
